requirements.txt
# Django packages
Django==5.1.*
django-debug-toolbar==4.4.*
django-environ==0.11.*
django-storages[s3]==1.14.*

django-crispy-forms==2.3
crispy_bootstrap5==2024.*
django-recaptcha==4.0.*

# WSGI HTTP Server
gunicorn==23.0.*

# Database packages
sqlparse==0.5.*
dj-database-url==2.2.*
psycopg[binary,pool]==3.2.*

# Media & static packages
pillow==10.4.*

# Mandatory packages
asgiref==3.8.*
pytz==2024.*
